--- Project/src/User/FogetPass.jsx
@@ -1,5 +1,0 @@
-function ForgetPass() {
-    return ( <></> );
-}
-
-export default ForgetPass;
