--- ds/áda.txt
@@ -1,1 +1,0 @@
-dấd
