+++ project-bck/src/constant.js
@@ -1,0 +1,13 @@
+//Trả về path trong trang web
+
+export const PATH_ROUTER = {
+    HOME: '/',
+    BEAR_CUTE: '/bear-cute',
+    BEAR_U: '/bear-U',
+    BEAR_BLANKETS: '/bear-blankets',
+    SUPPORT: '/support',
+    REGISTER: '/register',
+    LOGIN: '/login',
+    PAY: '/pay',
+    ACCOUNT: '/account'
+}
